requirements.txt
paho-mqtt==2.1.0
pydantic==2.12.5
